+++ src/postamble.js
@@ -212,8 +212,16 @@ Module['callMain'] = function callMain(args) {
     Module.realPrint('main() took ' + (Date.now() - start) + ' milliseconds');
 #endif
 
+#if EMTERPRETIFY_ASYNC
+    // if we are saving the stack, then do not call exit, we are not
+    // really exiting now, just unwinding the JS stack
+    if (EmterpreterAsync.state !== 1) {
+#endif // EMTERPRETIFY_ASYNC
     // if we're not running an evented main loop, it's time to exit
-    exit(ret, /* implicit = */ true);
+      exit(ret, /* implicit = */ true);
+#if EMTERPRETIFY_ASYNC
+    }
+#endif // EMTERPRETIFY_ASYNC
   }
   catch(e) {
     if (e instanceof ExitStatus) {
@@ -313,17 +321,62 @@ function run(args) {
 Module['run'] = run;
 
 function exit(status, implicit) {
-  if (implicit && Module['noExitRuntime']) {
 #if ASSERTIONS
-    Module.printErr('exit(' + status + ') implicitly called by end of main(), but noExitRuntime, so not exiting the runtime (you can use emscripten_force_exit, if you want to force a true shutdown)');
+#if NO_EXIT_RUNTIME == 1
+  // Compiler settings do not allow exiting the runtime, so flushing
+  // the streams is not possible. but in ASSERTIONS mode we check
+  // if there was something to flush, and if so tell the user they
+  // should request that the runtime be exitable.
+  // Normally we would not even include flush() at all, but in ASSERTIONS
+  // builds we do so just for this check, and here we see if there is any
+  // content to flush, that is, we check if there would have been
+  // something a non-ASSERTIONS build would have not seen.
+  // How we flush the streams depends on whether we are in NO_FILESYSTEM
+  // mode (which has its own special function for this; otherwise, all
+  // the code is inside libc)
+#if NO_FILESYSTEM
+  var flush = {{{ '$flush_NO_FILESYSTEM' in addedLibraryItems ? 'flush_NO_FILESYSTEM' : 'null' }}};
+#else
+  var flush = {{{ '$FS' in addedLibraryItems ? 'FS.quit' : "Module['_fflush']" }}};
 #endif
+  if (flush) {
+    var print = Module['print'];
+    var printErr = Module['printErr'];
+    var has = false;
+    Module['print'] = Module['printErr'] = function(x) {
+      has = true;
+    }
+    try { // it doesn't matter if it fails
+      flush(0);
+    } catch(e) {}
+    Module['print'] = print;
+    Module['printErr'] = printErr;
+    if (has) {
+      Runtime.warnOnce('stdio streams had content in them that was not flushed. you should set NO_EXIT_RUNTIME to 0 (see the FAQ), or make sure to emit a newline when you printf etc.');
+    }
+  }
+#endif // NO_EXIT_RUNTIME
+#endif // ASSERTIONS
+
+  // if this is just main exit-ing implicitly, and the status is 0, then we
+  // don't need to do anything here and can just leave. if the status is
+  // non-zero, though, then we need to report it.
+  // (we may have warned about this earlier, if a situation justifies doing so)
+  if (implicit && Module['noExitRuntime'] && status === 0) {
     return;
   }
 
   if (Module['noExitRuntime']) {
 #if ASSERTIONS
-    Module.printErr('exit(' + status + ') called, but noExitRuntime, so halting execution but not exiting the runtime or preventing further async execution (you can use emscripten_force_exit, if you want to force a true shutdown)');
-#endif
+    // if exit() was called, we may warn the user if the runtime isn't actually being shut down
+    if (!implicit) {
+#if NO_EXIT_RUNTIME
+      Module.printErr('exit(' + status + ') called, but NO_EXIT_RUNTIME is set, so halting execution but not exiting the runtime or preventing further async execution (build with NO_EXIT_RUNTIME=0, if you want a true shutdown)');
+#else
+      Module.printErr('exit(' + status + ') called, but noExitRuntime is set due to an async operation, so halting execution but not exiting the runtime or preventing further async execution (you can use emscripten_force_exit, if you want to force a true shutdown)');
+#endif // NO_EXIT_RUNTIME
+    }
+#endif // ASSERTIONS
   } else {
 #if USE_PTHREADS
     PThread.terminateAllThreads();